+++ src/main/java/org/rumbledb/runtime/flwor/clauses/LetClauseSparkIterator.java
@@ -36,7 +36,7 @@ import sparksoniq.jsoniq.ExecutionMode;
 import sparksoniq.jsoniq.tuple.FlworTuple;
 import sparksoniq.semantics.DynamicContext;
 import org.rumbledb.runtime.flwor.FlworDataFrameUtils;
-import sparksoniq.spark.udf.LetClauseUDF;
+import org.rumbledb.runtime.flwor.udfs.LetClauseUDF;
 
 import java.util.ArrayList;
 import java.util.Arrays;
