+++ src/main/java/org/rumbledb/runtime/flwor/clauses/OrderByClauseSparkIterator.java
@@ -45,8 +45,8 @@ import sparksoniq.jsoniq.tuple.FlworTuple;
 import sparksoniq.semantics.DynamicContext;
 import sparksoniq.semantics.types.ItemTypes;
 import org.rumbledb.runtime.flwor.FlworDataFrameUtils;
-import sparksoniq.spark.udf.OrderClauseCreateColumnsUDF;
-import sparksoniq.spark.udf.OrderClauseDetermineTypeUDF;
+import org.rumbledb.runtime.flwor.udfs.OrderClauseCreateColumnsUDF;
+import org.rumbledb.runtime.flwor.udfs.OrderClauseDetermineTypeUDF;
 
 import java.util.ArrayList;
 import java.util.HashSet;
