+++ src/main/java/org/rumbledb/runtime/flwor/clauses/WhereClauseSparkIterator.java
@@ -35,7 +35,7 @@ import sparksoniq.jsoniq.ExecutionMode;
 import sparksoniq.jsoniq.tuple.FlworTuple;
 import sparksoniq.semantics.DynamicContext;
 import org.rumbledb.runtime.flwor.FlworDataFrameUtils;
-import sparksoniq.spark.udf.WhereClauseUDF;
+import org.rumbledb.runtime.flwor.udfs.WhereClauseUDF;
 
 import java.util.HashSet;
 import java.util.List;
