+++ src/main/java/org/rumbledb/runtime/flwor/udfs/ForClauseUDF.java
@@ -18,12 +18,12 @@
  *
  */
 
-package sparksoniq.spark.udf;
+package org.rumbledb.runtime.flwor.udfs;
 
 import com.esotericsoftware.kryo.Kryo;
 import com.esotericsoftware.kryo.io.Input;
+import com.esotericsoftware.kryo.io.Output;
 import org.apache.spark.sql.api.java.UDF2;
-import org.apache.spark.sql.types.StructType;
 import org.rumbledb.api.Item;
 import org.rumbledb.items.ItemFactory;
 import org.rumbledb.runtime.RuntimeIterator;
@@ -37,49 +37,51 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
-public class WhereClauseUDF implements UDF2<WrappedArray<byte[]>, WrappedArray<Long>, Boolean> {
+public class ForClauseUDF implements UDF2<WrappedArray<byte[]>, WrappedArray<Long>, List<byte[]>> {
+    /**
+     *
+     */
     private static final long serialVersionUID = 1L;
-    private RuntimeIterator expression;
-    private Map<String, DynamicContext.VariableDependency> dependencies;
-
     private Map<String, List<String>> columnNamesByType;
-
+    private RuntimeIterator expression;
     private List<List<Item>> deserializedParams;
     private List<Item> longParams;
     private DynamicContext context;
-    private DynamicContext parentContext;
+    private List<Item> nextResult;
+    private List<byte[]> results;
 
     private transient Kryo kryo;
+    private transient Output output;
     private transient Input input;
 
-    public WhereClauseUDF(
+    public ForClauseUDF(
             RuntimeIterator expression,
             DynamicContext context,
-            StructType inputSchema,
             Map<String, List<String>> columnNamesByType
     ) {
         this.expression = expression;
+        this.columnNamesByType = columnNamesByType;
 
         this.deserializedParams = new ArrayList<>();
         this.longParams = new ArrayList<>();
-        this.parentContext = context;
-        this.context = new DynamicContext(this.parentContext);
+
+        this.context = new DynamicContext(context);
+        this.nextResult = new ArrayList<>();
+        this.results = new ArrayList<>();
 
         this.kryo = new Kryo();
         this.kryo.setReferences(false);
         FlworDataFrameUtils.registerKryoClassesKryo(this.kryo);
+        this.output = new Output(128, -1);
         this.input = new Input();
-
-        this.columnNamesByType = columnNamesByType;
-        this.dependencies = this.expression.getVariableDependencies();
-
     }
 
 
     @Override
-    public Boolean call(WrappedArray<byte[]> wrappedParameters, WrappedArray<Long> wrappedParametersLong) {
+    public List<byte[]> call(WrappedArray<byte[]> wrappedParameters, WrappedArray<Long> wrappedParametersLong) {
         this.deserializedParams.clear();
         this.context.removeAllVariables();
+        this.results.clear();
 
         FlworDataFrameUtils.deserializeWrappedParameters(
             wrappedParameters,
@@ -106,9 +108,15 @@ public class WhereClauseUDF implements UDF2<WrappedArray<byte[]>, WrappedArray<L
 
         // apply expression in the dynamic context
         this.expression.open(this.context);
-        boolean result = RuntimeIterator.getEffectiveBooleanValue(this.expression);
+        while (this.expression.hasNext()) {
+            this.nextResult.clear();
+            Item nextItem = this.expression.next();
+            this.nextResult.add(nextItem);
+            this.results.add(FlworDataFrameUtils.serializeItemList(this.nextResult, this.kryo, this.output));
+        }
         this.expression.close();
-        return result;
+
+        return this.results;
     }
 
     private void readObject(java.io.ObjectInputStream in)
@@ -119,6 +127,7 @@ public class WhereClauseUDF implements UDF2<WrappedArray<byte[]>, WrappedArray<L
         this.kryo = new Kryo();
         this.kryo.setReferences(false);
         FlworDataFrameUtils.registerKryoClassesKryo(this.kryo);
+        this.output = new Output(128, -1);
         this.input = new Input();
     }
 }
