+++ src/main/java/org/rumbledb/runtime/flwor/udfs/LetClauseUDF.java
@@ -18,7 +18,7 @@
  *
  */
 
-package sparksoniq.spark.udf;
+package org.rumbledb.runtime.flwor.udfs;
 
 import com.esotericsoftware.kryo.Kryo;
 import com.esotericsoftware.kryo.io.Input;
